+++ testcases/example.gyat
@@ -18,7 +18,11 @@ bop kai_cenat():
             duke_dennis = 10
             just put the fries in the bag bro
 
-    yap(duke_dennis)
+    hawk:
+        yap(duke_dennis)
+    tuah:
+        [baby_gronk rizz 10 mewing baby_gronk diddy huzz(69)]
+    
     its giving Aura
 
     hawk:
